// UI WebApp/ProjectTitanUICode/src/app/Interfaces/shipment.ts
export class Shipment {

    shipmentId: any;
    deviceId: any;
    productList: any;

    sourceLocation: any;
    destinationLocation: any;
    logisticsPartner: any;
    dateOfShipment: any;
    dateOfDelivery: any;
    purchaseOrderNumber: any;

    shipmentInfo: any;
    status: any;
    docRef: any;
    heartBeat: any;

    alertCount: any;
    alertDetails: any;
}